move animate-to.ts
